// persona/frontend/src/types/openapi.ts
export interface components {
    schemas: {
// ...
        BlogPostItem: {
            /** Id */
            id: string;
            /**
             * Status
             * @default pending
             */
            status: string;
// ...
        };
// ...
        TriggerResponse: {
            /** Ok */
            ok: boolean;
            /** Message */
            message: string;
            /** Label */
            label: string;
            /** Rate Limits */
            rate_limits?: {
                [key: string]: {
                    [key: string]: number;
                };
            } | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}